moves APIFeature class to separate file in routs/utils folder
